feat(employee): Add delete employee to repository, add delete route, delete employee ervice duplicate, add delete employee by id function, add delete employee controller
