+++ src/api/routes/employee.routes.ts
@@ -5,6 +5,8 @@ import { checkAuthToken } from '../middlewares/auth.middleware';
 const router = Router();
 
 router.post('/signup', checkAuthToken, EmployeeController.userExists);
-router.get('/', checkAuthToken, EmployeeController.getAllEmployees);
+router.get('/',  EmployeeController.getAllEmployees);
+router.delete('/delete/:id', EmployeeController.deleteEmployee);
+
 
 export { router as EmployeeRouter };
